requirements.txt
Django>=4.2
djangorestframework>=3.16
psutil>=5.9